Add local optimization for 5/355 models
