+++ TMIAutomation/Server/pyproject.toml
@@ -2,3 +2,6 @@
 name = "tmi-isocenter-server"
 version = "0.0.1"
 requires-python = ">=3.11"
+
+[tool.setuptools]
+packages = ["src"]
